Set up navigation from home page directory items
